// server/src/types/Catalog.ts
export interface AperturaCaja {
  codigo: number;
  fecha: string;
  monto: number;
  cajaCodigo: number;
  cajaDescripcion: string;
  usuarioCodigo: number;
  usuarioUsername: string;
  usuarioNombre: string;
  usuarioApellido: string;
  estadoCodigo: number;
  estadoDescripcion: string;
  fechaGrabacion: string;
  detalles: AperturaDetalle[];
  subtotal: number;
}

export interface AperturaDetalle {
  monedaCodigo: number;
  denominacion: string;
  tasa: number;
  cantidad: number;
  monto: number;
}

export interface AperturaDetalleInput {
  monedaCodigo: number;
  cantidad: number;
}
// ...
export interface AperturaCajaInput {
  cajaCodigo: number;
  estadoCodigo: number;
  detalles: AperturaDetalleInput[];
}